+++ vscode/CHANGELOG.md
@@ -35,6 +35,8 @@ This is a log of all notable changes to Cody for VS Code. [Unreleased] changes a
 
 ### Changed
 
+- The chat history is now associated to the currently logged in account. [issues/2261](https://github.com/sourcegraph/cody/issues/2261)
+
 ## [1.0.1]
 
 ### Added